+++ CHANGELOG.md
@@ -5,6 +5,8 @@
    - Skip unrecognized descriptors when importing archives rather than
      aborting the entire import.
    - Add new "host_name" parameter to filter by host name.
+   - Add new "unreachable_or_addresses" field with declared but
+     unreachable OR addresses.
 
 
 # Changes in version 4.2-1.6.1 - 2017-10-26